login alert 2번뜨는 오류 수정

+++ views/login/caller.js
@@ -1,10 +1,6 @@
 import AsyncStorage from '@react-native-community/async-storage';
 import React, {Component} from 'react';
-import {
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
+import { StyleSheet, Text, View, Alert } from 'react-native';
 import CustomButton from './custom_button';
 import { Item, Input} from 'native-base';
 import Icon from 'react-native-vector-icons/Ionicons';
@@ -42,28 +38,30 @@ class LoginScreen extends Component {
 
   
   makeRequest() {
-    if(userinfo.user.email == '')
-      alert("이메일을 입력해주세요")
-    if(userinfo.user.password == '')
-      alert("비밀번호를 입력해주세요")
-    api
-      .post('/users/sign_in', userinfo)
-      .then((response) => {
-        console.log(response.data.token);
-        AsyncStorage.setItem('token', response.data.token);
-        AsyncStorage.setItem('user_id', String(response.data.id));
-        AsyncStorage.setItem('myLocation', String(response.data.location_auth));
-        
-        if (String(response.data.location_auth) == "true") {// already has location
-          this.props.navigation.navigate('postIndex')
-        } else {
-          this.props.navigation.navigate('MyPage_Location')
-        }
-      })
-      .catch(function (error) {
-        console.log("login fail")
-        alert("가입하신 정보를 다시 확인해주세요")
-      });
+    if (userinfo.user.email == '')
+      Alert.alert("이메일을 입력해주세요")
+    if (userinfo.user.password == '')
+      Alert.alert("비밀번호를 입력해주세요")
+    if (!(userinfo.user.email == '') && !(userinfo.user.password == '')) {
+      api
+        .post('/users/sign_in', userinfo)
+        .then((response) => {
+          console.log(response.data.token);
+          AsyncStorage.setItem('token', response.data.token);
+          AsyncStorage.setItem('user_id', String(response.data.id));
+          AsyncStorage.setItem('myLocation', String(response.data.location_auth));
+
+          if (String(response.data.location_auth) == "true") {// already has location
+            this.props.navigation.navigate('postIndex')
+          } else {
+            this.props.navigation.navigate('MyPage_Location')
+          }
+        })
+        .catch(function (error) {
+          console.log("login fail")
+          alert("가입하신 정보를 다시 확인해주세요")
+        });
+    }
   }
 
   testLoginRequest(){
